requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.38.0
gunicorn==23.0.0
sqlalchemy==2.0.44
python-jose==3.5.0
bcrypt==5.0.0
python-multipart==0.0.20
email-validator==2.3.0
python-dotenv==1.1.1
chromadb==1.3.0
requests==2.31.0
pymssql==2.3.13
pysqlite3-binary